+++ Sources/BitReader.swift
@@ -35,6 +35,9 @@ public protocol BitReader: class {
     /// Reads `fromBits` bits and returns them as an `Int` number, advancing by `fromBits` BIT positions.
     func int(fromBits count: Int) -> Int
 
+    /// Reads `fromBits` bits and returns them as an `UInt8` number, advancing by `fromBits` BIT positions.
+    func byte(fromBits count: Int) -> UInt8
+
     /// Reads `fromBits` bits and returns them as an `UInt16` number, advancing by `fromBits` BIT positions.
     func uint16(fromBits count: Int) -> UInt16
 
@@ -61,10 +64,19 @@ public protocol BitReader: class {
     /// Reads 8 bytes and returns them as a `UInt64` number, advancing by 8 BYTE positions.
     func uint64() -> UInt64
 
+    /// Reads `fromBytes` bytes and returns them as a `UInt64` number, advancing by 8 BYTE positions.
+    func uint64(fromBytes count: Int) -> UInt64
+
     /// Reads 4 bytes and returns them as a `UInt32` number, advancing by 4 BYTE positions.
     func uint32() -> UInt32
 
+    /// Reads `fromBytes` bytes and returns them as a `UInt32` number, advancing by 8 BYTE positions.
+    func uint32(fromBytes count: Int) -> UInt32
+
     /// Reads 2 bytes and returns them as a `UInt16` number, advancing by 2 BYTE positions.
     func uint16() -> UInt16
+
+    /// Reads `fromBytes` bytes and returns them as a `UInt16` number, advancing by 8 BYTE positions.
+    func uint16(fromBytes count: Int) -> UInt16
 
 }