+++ Sources/ByteReader.swift
@@ -22,9 +22,18 @@ public class ByteReader {
 
      - Note: It generally means that all bytes have been read.
      */
-
     public var isFinished: Bool {
         return self.data.endIndex <= self.offset
+    }
+
+    /// Amount of bytes left to read.
+    public var bytesLeft: Int {
+        return self.data.endIndex - self.offset
+    }
+
+    /// Amount of bytes that were already read.
+    public var bytesRead: Int {
+        return self.offset - self.data.startIndex
     }
 
     /// Creates an instance for reading bytes from `data`.
@@ -53,9 +62,7 @@ public class ByteReader {
      */
     public func bytes(count: Int) -> [UInt8] {
         precondition(count >= 0)
-        guard count > 0
-            else { return [] }
-        precondition(self.offset + count <= self.data.endIndex)
+        precondition(bytesLeft >= count)
         defer { self.offset += count }
         return self.data[self.offset..<self.offset + count].toArray(type: UInt8.self, count: count)
     }
@@ -68,9 +75,7 @@ public class ByteReader {
      */
     public func int(fromBytes count: Int) -> Int {
         precondition(count >= 0)
-        guard count > 0
-            else { return 0 }
-        precondition(self.offset + count <= self.data.endIndex)
+        precondition(bytesLeft >= count)
         // TODO: If uintX() could be force inlined or something in the future than probably it would make sense
         // to use them for `count` == 2, 4 or 8.
         var result = 0
@@ -87,9 +92,29 @@ public class ByteReader {
      - Precondition: There MUST be enough data left.
      */
     public func uint64() -> UInt64 {
-        precondition(self.offset + 8 <= self.data.endIndex)
+        precondition(bytesLeft >= 8)
         defer { self.offset += 8 }
         return self.data[self.offset..<self.offset + 8].to(type: UInt64.self)
+    }
+
+    /**
+     Reads `fromBytes` bytes and returns them as an `UInt64` number, advancing by `fromBytes` positions.
+
+     - Note: If it is known that `fromBytes` is exactly 8, then consider using `uint64()` function (without argument),
+     since it has better performance in this situation.
+     - Precondition: Parameter `fromBits` MUST be from `0..8` range, i.e. it MUST not exceed maximum possible amount of
+     bytes that `UInt64` type can represent.
+     - Precondition: There MUST be enough data left.
+     */
+    public func uint64(fromBytes count: Int) -> UInt64 {
+        precondition(0...8 ~= count)
+        precondition(bytesLeft >= count)
+        var result = 0 as UInt64
+        for i in 0..<count {
+            result += UInt64(truncatingIfNeeded: self.data[self.offset]) << (8 * i)
+            self.offset += 1
+        }
+        return result
     }
 
     /**
@@ -98,9 +123,29 @@ public class ByteReader {
      - Precondition: There MUST be enough data left.
      */
     public func uint32() -> UInt32 {
-        precondition(self.offset + 4 <= self.data.endIndex)
+        precondition(bytesLeft >= 4)
         defer { self.offset += 4 }
         return self.data[self.offset..<self.offset + 4].to(type: UInt32.self)
+    }
+
+    /**
+     Reads `fromBytes` bytes and returns them as an `UInt32` number, advancing by `fromBytes` positions.
+
+     - Note: If it is known that `fromBytes` is exactly 4, then consider using `uint32()` function (without argument),
+     since it has better performance in this situation.
+     - Precondition: Parameter `fromBits` MUST be from `0..4` range, i.e. it MUST not exceed maximum possible amount of
+     bytes that `UInt32` type can represent.
+     - Precondition: There MUST be enough data left.
+     */
+    public func uint32(fromBytes count: Int) -> UInt32 {
+        precondition(0...4 ~= count)
+        precondition(bytesLeft >= count)
+        var result = 0 as UInt32
+        for i in 0..<count {
+            result += UInt32(truncatingIfNeeded: self.data[self.offset]) << (8 * i)
+            self.offset += 1
+        }
+        return result
     }
 
     /**
@@ -109,9 +154,29 @@ public class ByteReader {
      - Precondition: There MUST be enough data left.
      */
     public func uint16() -> UInt16 {
-        precondition(self.offset + 2 <= self.data.endIndex)
+        precondition(bytesLeft >= 2)
         defer { self.offset += 2 }
         return self.data[self.offset..<self.offset + 2].to(type: UInt16.self)
+    }
+
+    /**
+     Reads `fromBytes` bytes and returns them as an `UInt16` number, advancing by `fromBytes` positions.
+
+     - Note: If it is known that `fromBytes` is exactly 2, then consider using `uint16()` function (without argument),
+     since it has better performance in this situation.
+     - Precondition: Parameter `fromBits` MUST be from `0..2` range, i.e. it MUST not exceed maximum possible amount of
+     bytes that `UInt16` type can represent.
+     - Precondition: There MUST be enough data left.
+     */
+    public func uint16(fromBytes count: Int) -> UInt16 {
+        precondition(0...2 ~= count)
+        precondition(bytesLeft >= count)
+        var result = 0 as UInt16
+        for i in 0..<count {
+            result += UInt16(truncatingIfNeeded: self.data[self.offset]) << (8 * i)
+            self.offset += 1
+        }
+        return result
     }
 
 }