+++ Sources/MsbBitReader.swift
@@ -18,21 +18,12 @@ public final class MsbBitReader: ByteReader, BitReader {
 
     /// Amount of bits left to read.
     public var bitsLeft: Int {
-        if self.isFinished {
-            return 0
-        } else {
-            return (self.data.endIndex - self.offset) * 8 - self.bitMask.leadingZeroBitCount
-        }
+        return self.bytesLeft * 8 - self.bitMask.leadingZeroBitCount
     }
 
     /// Amount of bits that were already read.
     public var bitsRead: Int {
-        if self.isFinished {
-            return 8 * self.size
-        } else {
-            return (self.offset - self.data.startIndex) * 8 + self.bitMask.leadingZeroBitCount
-        }
-
+        return self.bytesRead * 8 + self.bitMask.leadingZeroBitCount
     }
 
     /// Creates an instance for reading bits (and bytes) from `data`.
@@ -46,7 +37,7 @@ public final class MsbBitReader: ByteReader, BitReader {
      `byteReader` is preserved.
      */
     public init(_ byteReader: ByteReader) {
-        self.currentByte = byteReader.offset < byteReader.data.endIndex ? byteReader.data[byteReader.offset] : 0
+        self.currentByte = byteReader.isFinished ? 0 : byteReader.data[byteReader.offset]
         super.init(data: byteReader.data)
         self.offset = byteReader.offset
     }
@@ -78,8 +69,6 @@ public final class MsbBitReader: ByteReader, BitReader {
      */
     public func bits(count: Int) -> [UInt8] {
         precondition(count >= 0)
-        guard count > 0
-            else { return [] }
         precondition(bitsLeft >= count)
 
         var array = [UInt8]()
@@ -94,14 +83,12 @@ public final class MsbBitReader: ByteReader, BitReader {
     /**
      Reads `fromBits` bits and returns them as an `Int` number, advancing by `fromBits` BIT positions.
 
-     - Precondition: Parameter `fromBits` MUST be from `0..<Int.bitWidth` range, i.e. it MUST not exceed maximum bit
+     - Precondition: Parameter `fromBits` MUST be from `0...Int.bitWidth` range, i.e. it MUST not exceed maximum bit
      width on the current platform.
      - Precondition: There MUST be enough data left.
      */
     public func int(fromBits count: Int) -> Int {
-        precondition(0..<Int.bitWidth ~= count)
-        guard count > 0
-            else { return 0 }
+        precondition(0...Int.bitWidth ~= count)
         precondition(bitsLeft >= count)
 
         var result = 0
@@ -121,15 +108,41 @@ public final class MsbBitReader: ByteReader, BitReader {
     }
 
     /**
+     Reads `fromBits` bits and returns them as an `UInt8` number, advancing by `fromBits` BIT positions.
+
+     - Precondition: Parameter `fromBits` MUST be from `0...8` range, i.e. it MUST not exceed maximum bit width of
+     `UInt8` type on the current platform.
+     - Precondition: There MUST be enough data left.
+     */
+    public func byte(fromBits count: Int) -> UInt8 {
+        precondition(0...8 ~= count)
+        precondition(bitsLeft >= count)
+
+        var result = 0 as UInt8
+        for i in 0..<count {
+            let bit: UInt8 = self.currentByte & self.bitMask > 0 ? 1 : 0
+            result += (1 << (count - i - 1)) * bit
+
+            if self.bitMask == 1 {
+                self.offset += 1
+                self.bitMask = 128
+            } else {
+                self.bitMask >>= 1
+            }
+        }
+
+        return result
+    }
+
+    /**
      Reads `fromBits` bits and returns them as an `UInt16` number, advancing by `fromBits` BIT positions.
 
-     - Precondition: Parameter `fromBits` MUST be from `0..<16` range.
+     - Precondition: Parameter `fromBits` MUST be from `0...16` range, i.e. it MUST not exceed maximum bit width of
+     `UInt16` type on the current platform.
      - Precondition: There MUST be enough data left.
      */
     public func uint16(fromBits count: Int) -> UInt16 {
-        precondition(0..<16 ~= count)
-        guard count > 0
-            else { return 0 }
+        precondition(0...16 ~= count)
         precondition(bitsLeft >= count)
 
         var result = 0 as UInt16
@@ -151,13 +164,12 @@ public final class MsbBitReader: ByteReader, BitReader {
     /**
      Reads `fromBits` bits and returns them as an `UInt32` number, advancing by `fromBits` BIT positions.
 
-     - Precondition: Parameter `fromBits` MUST be from `0..<32` range.
+     - Precondition: Parameter `fromBits` MUST be from `0...32` range, i.e. it MUST not exceed maximum bit width of
+     `UInt32` type on the current platform.
      - Precondition: There MUST be enough data left.
      */
     public func uint32(fromBits count: Int) -> UInt32 {
-        precondition(0..<32 ~= count)
-        guard count > 0
-            else { return 0 }
+        precondition(0...32 ~= count)
         precondition(bitsLeft >= count)
 
         var result = 0 as UInt32
@@ -179,13 +191,12 @@ public final class MsbBitReader: ByteReader, BitReader {
     /**
      Reads `fromBits` bits and returns them as an `UInt64` number, advancing by `fromBits` BIT positions.
 
-     - Precondition: Parameter `fromBits` MUST be from `0..<64` range.
+     - Precondition: Parameter `fromBits` MUST be from `0...64` range, i.e. it MUST not exceed maximum bit width of
+     `UInt64` type on the current platform.
      - Precondition: There MUST be enough data left.
      */
     public func uint64(fromBits count: Int) -> UInt64 {
-        precondition(0..<64 ~= count)
-        guard count > 0
-            else { return 0 }
+        precondition(0...64 ~= count)
         precondition(bitsLeft >= count)
 
         var result = 0 as UInt64
@@ -221,7 +232,11 @@ public final class MsbBitReader: ByteReader, BitReader {
 
     // MARK: ByteReader's methods.
 
-    /// Offset to the byte in `data` which will be read next.
+    /**
+     Offset to the byte in `data` which will be read next.
+
+     - Note: The byte which is currently used for reading bits from is included into `bytesRead`.
+     */
     public override var offset: Int {
         didSet {
             if !self.isFinished {
@@ -276,6 +291,18 @@ public final class MsbBitReader: ByteReader, BitReader {
     }
 
     /**
+     Reads `fromBytes` bytes and returns them as a `UInt64` number, advancing by `fromBytes` BYTE positions.
+
+     - Precondition: Reader MUST be aligned.
+     - Precondition: Parameter `fromBytes` MUST not be less than 0.
+     - Precondition: There MUST be enough data left.
+     */
+    public override func uint64(fromBytes count: Int) -> UInt64 {
+        precondition(isAligned, "BitReader is not aligned.")
+        return super.uint64(fromBytes: count)
+    }
+
+    /**
      Reads 4 bytes and returns them as a `UInt32` number, advancing by 4 BYTE positions.
 
      - Precondition: Reader MUST be aligned.
@@ -287,6 +314,18 @@ public final class MsbBitReader: ByteReader, BitReader {
     }
 
     /**
+     Reads `fromBytes` bytes and returns them as a `UInt32` number, advancing by `fromBytes` BYTE positions.
+
+     - Precondition: Reader MUST be aligned.
+     - Precondition: Parameter `fromBytes` MUST not be less than 0.
+     - Precondition: There MUST be enough data left.
+     */
+    public override func uint32(fromBytes count: Int) -> UInt32 {
+        precondition(isAligned, "BitReader is not aligned.")
+        return super.uint32(fromBytes: count)
+    }
+
+    /**
      Reads 2 bytes and returns them as a `UInt16` number, advancing by 2 BYTE positions.
 
      - Precondition: Reader MUST be aligned.
@@ -295,6 +334,18 @@ public final class MsbBitReader: ByteReader, BitReader {
     public override func uint16() -> UInt16 {
         precondition(isAligned, "BitReader is not aligned.")
         return super.uint16()
+    }
+
+    /**
+     Reads `fromBytes` bytes and returns them as a `UInt16` number, advancing by `fromBytes` BYTE positions.
+
+     - Precondition: Reader MUST be aligned.
+     - Precondition: Parameter `fromBytes` MUST not be less than 0.
+     - Precondition: There MUST be enough data left.
+     */
+    public override func uint16(fromBytes count: Int) -> UInt16 {
+        precondition(isAligned, "BitReader is not aligned.")
+        return super.uint16(fromBytes: count)
     }
 
 }